Refactor plugins and injection locations
