+++ packages/slate-react/src/plugins/dom/index.js
@@ -1,5 +1,8 @@
 import { IS_ANDROID } from 'slate-dev-environment'
+
 import AndroidPlugin from '../android'
+import NoopPlugin from '../debug/noop'
+
 import AfterPlugin from './after'
 import BeforePlugin from './before'
 
@@ -18,9 +21,11 @@ function DOMPlugin(options = {}) {
   // COMPAT: Add Android specific handling separately before it gets to the
   // other plugins because it is specific (other browser don't need it) and
   // finicky (it has to come before other plugins to work).
-  const beforeBeforePlugins = IS_ANDROID ? [AndroidPlugin()] : []
+  const androidPlugins = IS_ANDROID
+    ? [AndroidPlugin(options), NoopPlugin(options)]
+    : []
 
-  return [...beforeBeforePlugins, beforePlugin, ...plugins, afterPlugin]
+  return [...androidPlugins, beforePlugin, ...plugins, afterPlugin]
 }
 
 /**
